requirements.txt
flask>=2.0.0
langchain-groq>=0.1.0
python-dotenv>=0.19.0
groq>=0.3.0